sync (#63)
* add sagemaker cli (#32)

* add sagemaker cli

* remove unnecessary close

* address PR comments

* tidy up imports

* fix imports, flake8 errors

* improve help message for bucket-name

* remove default role name

* fix log-level and py3 tests, add copyright

* update cli example scripts

* Add documentation about BYO Models (#47)

* Add test for BYO estimator using Factorization Machines algorithm as an example. (#50)

* Support multi-part uploads (#45)

* Update TensorFlow examples following API change (#44)

* Add data_type to hyperparameters (#54)

When we describe a training job the data type of the hyper parameters is
lost because we use a dict[str, str]. This adds a new field to
Hyperparameter so that we can convert the datatypes at runtime.

instead of validating with isinstance(), we cast the hp value to the type it
is meant to be. This enforces a "strongly typed" value. When we
deserialize from the API string responses it becomes easier to deal with
too.