Add TF CIFAR test to continuous testing set
Though this test has historically been relatively flaky, there was
a recent change to the training steps that helped stabilize it, so
it should be fine to include this test with our continuous testing.
This test is important because it's currently our only one that tests
distributed training and the only one to run on GPU instances.