Refactor LocalSageMakerClient
local sagemakerclient was not very useful beyond creating a single
estimator/endpoint. doing workflows such as training and later creating
and endpoint was really awkward.

These changes make it both resemble the API a bit more and allow
persisting objects across LocalSessions. This is important because most
of the time the SDK classes create sessions behind the scenes.