Document upcoming MXNet training script format (#390)
The next major release of MXNet will change the
training script format. This README change documents the
changes needed by the user to adjust to the new format.
This is currently just a warning as the new format is not out yet.
The warning is meant to help users plan for the upcoming change.