Add Local Mode Batch Transform support. (#414)
Creating local batch transform jobs will perform all the steps locally. The serving container will be created the same way as it is for serving but then the data will be fed to it and written to the corresponding output location. Data splits and batch strategies are implemented as closely as possible to the behavior in SageMaker so that the end results are consistent.