change: eliminate dependency on mnist dataset website
This commit will modify the folder structure to avoid needing to
download the datasets and relying on the dataset website from being
available.
This caused a canary to fail (TT-0407832742).

The dependency has existed since Pytorch 1.1 upgrade, as the directory
structure was modified in https://github.com/pytorch/vision/pull/601