infra: explicitly check for test bucket before creation
Instead of relying on an exception thrown during creation, we now
explicitly test for a bucket's existence in one of the integration
tests.
This avoids a situation where we get a resource allocation exception
and unnecessarily fail.
Specifically: botocore.exceptions.ClientError: An error occurred
(OperationAborted) when calling the CreateBucket operation:
A conflicting conditional operation is currently in progress against
this resource. Please try again.