modify session and kms_utils to check for S3 bucket before creation
Previously, the code would create the bucket and enact business logic
based on the exceptions thrown.
This commit makes it such that the code checks that the bucket exists
before trying to create it.

In addition to being cleaner, this also avoids issues if customers
have no S3 bucket creation permissions.