fix: fix HyperparameterTuner.attach for Marketplace algorithms
The AlgorithmSpecification for a training job that is part of
a hyperparameter tuning job will have TrainingImage set to "",
rather than not being present in the training job description.
Looking for AlgorithmName first, instead of TrainingImage,
fixes the issue of algorithm_arn not being populated when
attaching a hyperparameter tuning job.