infra: add PyTorch + custom model bucket batch transform integ test
This converts the MXNet + VPC config + batch transform integ test to
use PyTorch instead, and start from a model rather than an estimator.
In addition, the model is uploaded to a non-default bucket, and
the tests checks that the repacked model is saved in the same
(non-default) bucket. The motivation for converting an existing test
is to not increase the overall runtime of the integ tests.