fix: ParamValidationError when scheduling a Clarify model monitor
For ModelExplainabilityMonitor, if the model_scores parameter of the method
suggest_baseline is an integer, the consequent create_monitoring_schedule
will fail with ParamValidationError. The rootcause is that the parameter is
passed to endpoint_input.inference_attribute which only accepts string.
The same issue raises for ModelBiasMonitor, if model_predicted_label_config
has integer label or probability which is passed to endpoint_input as well.
The commit fixes the issue by proper type conversion from int to str.