fix: remove DefaultEnumMeta used in the workflow package
The DefaultEnumMeta metaclass is malfunctioned. Enum types
defined with this couldn't be instantiated correctly.
For example, StepTypeEnum('Processing') leads to StepTypeEnum.CONDITION

Removed this metaclass altogether instead of fixing it because
the target behavior "defaulting to the first value in the Enum list"
is not particularly useful in the pipeline SDK implementation
and can be misleading even implemented correctly.