fix: add environment key to model container only if it is not empty
This caused a problem when trying to deploy a model package from model
registry into an endpoint. No environment variables were provided to the
model package, but it was registered as an empty mapping. When trying to
deploy the model package as an endpoint and providing some environment
variables to the endpoint, there was an error from the API:
Environment variable map cannot be specified both at Model Package {} and
Model {...}.

Fixes: aws/sagemaker-python-sdk#3397